% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chromato_line.R
\name{add_chromato_line}
\alias{add_chromato_line}
\title{Add chromato line}
\usage{
add_chromato_line(
  plot,
  chromato,
  shift = 0,
  time,
  name,
  color,
  polarity = "pos"
)
}
\arguments{
\item{plot}{Plot}

\item{chromato}{Chromato}

\item{shift}{Shift}

\item{time}{Time}

\item{name}{Name}

\item{color}{Color}

\item{polarity}{Polarity}
}
\value{
A plot with added chromato line
}
\description{
Add chromato line
}
\examples{
NULL
}
