% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peak_metadata.R
\name{add_peak_metadata}
\alias{add_peak_metadata}
\title{Add peak metadata}
\usage{
add_peak_metadata(df, min_score = 0.8)
}
\arguments{
\item{df}{Dataframe}

\item{min_score}{Min score}
}
\value{
A dataframe with peak metadata
}
\description{
Add peak metadata
}
\examples{
NULL
}
