% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mods_colors.R
\name{assign_mods_colors}
\alias{assign_mods_colors}
\title{Assign index cascade colors (WIP)}
\usage{
assign_mods_colors(.data)
}
\arguments{
\item{.data}{an msd based dataframe}
}
\value{
a color encoded data frame of modalities
}
\description{
Assign index cascade colors (WIP)
}
