% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_cascade.R
\name{batch_plot}
\alias{batch_plot}
\title{Title}
\usage{
batch_plot(
  msd_df,
  cscd_num,
  export = TRUE,
  imgtype = ".png",
  imgpath = "Images"
)
}
\arguments{
\item{msd_df}{msd based data frame to be passed through cascade call}

\item{cscd_num}{number from the cascade plot see \code{plot_name} for a crosswalk}

\item{export}{logical to control saving of plot, TRUE by default}

\item{imgtype}{type of image to save (default = .png, .svg, .jpeg,)}

\item{imgpath}{path where images are saved ("Images" set as default)}
}
\value{
a ggplot2 object
}
\description{
Title
}
\examples{
\dontrun{
plt <- batch_plot(df_sub, cscd_num = 1, export = T, imgtype = ".png", imgpath = "Images")}


}
