% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_munging.R
\name{bind_tx_curr_lag}
\alias{bind_tx_curr_lag}
\title{Title}
\usage{
bind_tx_curr_lag(df, df2)
}
\arguments{
\item{df}{base data frame to bind to}

\item{df2}{tx_curr_lag2 data frame to bind on}
}
\value{
a single bound data frame
}
\description{
Title
}
