% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{CAQTL_STATUS}
\alias{CAQTL_STATUS}
\title{caQTL Status Descriptions}
\format{
An object of class \code{list} of length 3.
}
\usage{
CAQTL_STATUS
}
\description{
caQTL Status Descriptions
}
\keyword{datasets}
