% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cascade_cache_clear}
\alias{cascade_cache_clear}
\title{Clear CASCADE Cache}
\usage{
cascade_cache_clear(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Remove all cached data for a configuration
}
