% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cascade_cache_info}
\alias{cascade_cache_info}
\title{Get CASCADE Cache Information}
\usage{
cascade_cache_info(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Display cache statistics and information
}
