% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{cascade_derivs}
\alias{cascade_derivs}
\title{Compute derivatives for cascade model}
\usage{
cascade_derivs(y, p)
}
\arguments{
\item{y}{initial.}

\item{p}{parameters.}
}
\description{
Compute derivatives for cascade model.
}

