% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_intensity_name.R
\name{change_intensity_name}
\alias{change_intensity_name}
\title{Change intensity name}
\usage{
change_intensity_name(df, name_rt = "rtime", name_intensity = "intensity")
}
\arguments{
\item{df}{Dataframe}

\item{name}{Name}
}
\value{
A dataframe with changed intensity name
}
\description{
Change intensity name
}
\examples{
NULL
}
