% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_load_packages.R
\name{check_and_load_packages_2}
\alias{check_and_load_packages_2}
\title{Check and load packages 2}
\usage{
check_and_load_packages_2(
  cran = packages_cran,
  bioconductor = packages_bioconductor,
  github = packages_github
)
}
\arguments{
\item{cran}{CRAN}

\item{bioconductor}{Bioconductor}

\item{github}{Github}
}
\value{
Log of installed packages
}
\description{
Check and load packages 2
}
\examples{
NULL
}
