% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_export_dir.R
\name{check_export_dir}
\alias{check_export_dir}
\title{Check export dir}
\usage{
check_export_dir(dir)
}
\arguments{
\item{dir}{Dir}
}
\value{
A log of checked dir
}
\description{
Check export dir
}
\examples{
NULL
}
