% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_checks.R
\name{check_metadata}
\alias{check_metadata}
\title{Check for metadata list}
\usage{
check_metadata()
}
\value{
message if list does not exist
}
\description{
Check for metadata list
}
\examples{
\dontrun{
check_metadata()
}
}
