% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peaks_integration.R
\name{check_peaks_integration}
\alias{check_peaks_integration}
\title{Check chromatograms alignment}
\usage{
check_peaks_integration(
  file = NULL,
  features = NULL,
  chromatogram = "baselined",
  min_area = 0.005,
  min_intensity = 10000,
  time_min = 0.7,
  time_max = 35.2,
  cad_shift = 0.05,
  show_example = FALSE
)
}
\arguments{
\item{file}{File path}

\item{features}{Features path}

\item{chromatogram}{Chromatogram}

\item{min_area}{Minimum area}

\item{min_intensity}{Minimum intensity}

\item{time_min}{Minimum time}

\item{time_max}{Maximum time}

\item{cad_shift}{CAD shift}

\item{show_example}{Show example? Default to FALSE}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_peaks_integration(show_example = TRUE)
}
}
