% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_checks.R
\name{check_plot_entry}
\alias{check_plot_entry}
\title{Check plot entry value
Validate plot entry from user}
\usage{
check_plot_entry(cscd_val)
}
\arguments{
\item{cscd_val}{number entered by user, from 1:13 for valid entries}
}
\value{
a message for invalid plot entries
}
\description{
Check plot entry value
Validate plot entry from user
}
\examples{
\dontrun{
check_plot_entry()
}


}
