% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache_internal}
\alias{clear_cache_internal}
\title{Clear Cache Internal}
\usage{
clear_cache_internal(cache_dir)
}
\arguments{
\item{cache_dir}{Cache directory}
}
\description{
Remove all cache files in directory
}
\keyword{internal}
