% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_susie.R
\name{combine_chromosome_results}
\alias{combine_chromosome_results}
\title{Combine SuSiE Results from Multiple Chromosomes}
\usage{
combine_chromosome_results(chr_results_list, cell_types)
}
\arguments{
\item{chr_results_list}{List of chromosome results}

\item{cell_types}{Cell type names}
}
\value{
Combined results structure with all chromosomes merged
}
\description{
Combines chromosome results from either parallel or sequential processing
}
\keyword{internal}
