% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_variants.R
\name{combine_chromosome_variant_data}
\alias{combine_chromosome_variant_data}
\title{Combine Variant Data Across Chromosomes}
\usage{
combine_chromosome_variant_data(chr_results)
}
\arguments{
\item{chr_results}{List of variant data by chromosome}
}
\value{
Combined variant data
}
\description{
Combine Variant Data Across Chromosomes
}
\keyword{internal}
