% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.R
\name{compare_peaks}
\alias{compare_peaks}
\title{Compare peaks}
\usage{
compare_peaks(x, list_ms_peaks, peaks_prelist)
}
\arguments{
\item{x}{X}

\item{list_ms_peaks}{list_ms_peaks}

\item{peaks_prelist}{peaks_prelist}
}
\value{
A comparison score
}
\description{
Compare peaks
}
\examples{
NULL
}
