% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{CONFIDENCE_SCORE_MIN}
\alias{CONFIDENCE_SCORE_MIN}
\title{Parameters for annotation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CONFIDENCE_SCORE_MIN
}
\description{
Parameters for annotation
}
\keyword{datasets}
