% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\docType{data}
\name{DEFAULT_CELL_HIERARCHY}
\alias{DEFAULT_CELL_HIERARCHY}
\title{Default Cell Type Hierarchy (Immune)}
\format{
An object of class \code{CellTypeHierarchy} of length 12.
}
\usage{
DEFAULT_CELL_HIERARCHY
}
\description{
The default hierarchy for immune cell QTL analysis, with myeloid/lymphoid
lineages and T-cell subgroup. Produces 6 specificity categories.
}
\keyword{datasets}
