% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{DEFAULT_COLUMN_MAPPING}
\alias{DEFAULT_COLUMN_MAPPING}
\title{Default Column Mappings}
\format{
An object of class \code{list} of length 9.
}
\usage{
DEFAULT_COLUMN_MAPPING
}
\description{
Maps internal column names to input column names for each file type.
Entries with NULL values are optional and skipped during rename.
Users can override individual entries via create_config(column_mapping = list(...)).
}
\keyword{datasets}
