% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim-methods}
\alias{dim,micro_array-method}
\title{Dimension of the data}
\usage{
\S4method{dim}{micro_array}(x)
}
\arguments{
\item{x}{an object of class "micro-array}
}
\description{
Dimension of the data
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"micro_array\")")}{ Gives the dimension of the
matrix of measurements. } }
}

\examples{

 if(require(CascadeData)){
	data(micro_US)
	micro_US<-as.micro_array(micro_US,time=c(60,90,210,390),subject=6)
	dim(micro_US)
	}

}
\keyword{methods}
