% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{EQTL_STATUS}
\alias{EQTL_STATUS}
\title{eQTL Status Descriptions}
\format{
An object of class \code{list} of length 4.
}
\usage{
EQTL_STATUS
}
\description{
eQTL Status Descriptions
}
\keyword{datasets}
