% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chromatogram.R
\name{extract_chromatogram}
\alias{extract_chromatogram}
\title{Extract chromatogram}
\usage{
extract_chromatogram(list, type)
}
\arguments{
\item{list}{List}

\item{type}{Type}
}
\value{
An extracted chromatogram
}
\description{
Extract chromatogram
}
\examples{
NULL
}
