% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{extract_variant_lfsr}
\alias{extract_variant_lfsr}
\title{Extract LFSR values for variants}
\usage{
extract_variant_lfsr(variant_matrix, lfsr_data, cell_types, feature_col)
}
\arguments{
\item{variant_matrix}{Variant matrix with QTL effects}

\item{lfsr_data}{LFSR data table}

\item{cell_types}{Vector of cell type column names}

\item{feature_col}{Column containing affected features}
}
\value{
data.table with LFSR values
}
\description{
Helper function to extract LFSR values for variant-feature pairs
}
\keyword{internal}
