% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msd_meta.R
\name{file_fy}
\alias{file_fy}
\title{Returns fiscal year information for file path of msd}
\usage{
file_fy(file_path)
}
\arguments{
\item{file_path}{path where msd lives}
}
\value{
current fiscal year used as a filter
}
\description{
Returns fiscal year information for file path of msd
}
