% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msd_meta.R
\name{file_pd}
\alias{file_pd}
\title{Returns period information from file path of msd}
\usage{
file_pd(file_path)
}
\arguments{
\item{file_path}{path where msd lives}
}
\value{
current period used as a filter
}
\description{
Returns period information from file path of msd
}
