% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msd_meta.R
\name{file_source}
\alias{file_source}
\title{Returns source of msd}
\usage{
file_source(file_path)
}
\arguments{
\item{file_path}{path where msd lives}
}
\value{
object with msd source
}
\description{
Returns source of msd
}
