% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fft.R
\name{filter_fft}
\alias{filter_fft}
\title{Filter FFT}
\usage{
filter_fft(x, components)
}
\arguments{
\item{x}{X}

\item{components}{Components}
}
\value{
The fourier filtered x
}
\description{
Filter FFT
}
\examples{
NULL
}
