% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{filter_l1_celltypes}
\alias{filter_l1_celltypes}
\title{Filter to only L1 cell types}
\usage{
filter_l1_celltypes(celltypes, hierarchy)
}
\arguments{
\item{celltypes}{Character vector of cell type column names}

\item{hierarchy}{A CellTypeHierarchy object (required)}
}
\value{
Character vector of L1-only cell types
}
\description{
Filter to only L1 cell types
}
