% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_munging.R
\name{fltr_ayp}
\alias{fltr_ayp}
\title{Filter for Adults and Young People}
\usage{
fltr_ayp(.data)
}
\arguments{
\item{.data}{MER structured data set whittled down}
}
\value{
passes through a filtered ayp dataframe
}
\description{
Adds a category to trendscoarse if indicating AYP or Non AYP
}
