% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_cascade.R
\name{fltr_cascade}
\alias{fltr_cascade}
\title{Cascade filter
Filters the cascade variables to current fiscal year}
\usage{
fltr_cascade(.data)
}
\arguments{
\item{.data}{MER structured data set}
}
\value{
passed through a filtered cascade data frame
}
\description{
Cascade filter
Filters the cascade variables to current fiscal year
}
