% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msd_meta.R
\name{footer_meta}
\alias{footer_meta}
\title{Returns ggplot caption information for file path of msd}
\usage{
footer_meta(file_path)
}
\arguments{
\item{file_path}{path where msd lives}
}
\value{
caption to be used in footnote
}
\description{
Returns ggplot caption information for file path of msd
}
