% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt.R
\name{format_gt}
\alias{format_gt}
\title{Temp GT function}
\usage{
format_gt(table, title = "", subtitle = "")
}
\arguments{
\item{table}{Table}

\item{title}{Title}

\item{subtitle}{Subtitle}
}
\value{
A formatted GT table
}
\description{
Temp GT function
}
\examples{
NULL
}
