% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{FOURRIER_COMPONENTS}
\alias{FOURRIER_COMPONENTS}
\title{Parameters for signal improvement}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FOURRIER_COMPONENTS
}
\description{
Parameters for signal improvement
}
\keyword{datasets}
