% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{generate_config_hash}
\alias{generate_config_hash}
\title{Generate Configuration Hash}
\usage{
generate_config_hash(config)
}
\arguments{
\item{config}{Configuration list}
}
\value{
Character hash string
}
\description{
Create hash of configuration elements that affect data loading
}
\keyword{internal}
