% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudochromatograms.R
\name{generate_pseudochromatograms}
\alias{generate_pseudochromatograms}
\title{Generate pseudochromatograms}
\usage{
generate_pseudochromatograms(
  annotations = NULL,
  features_informed = NULL,
  features_not_informed = NULL,
  file = NULL,
  headers = c(bpi = "BasePeak_0", pda = "PDA#1_TotalAbsorbance_0", cad = "UV#1_CAD_1_0"),
  detector = "cad",
  show_example = FALSE,
  min_confidence = 0.4,
  min_similarity_prefilter = 0.6,
  min_similarity_filter = 0.8,
  mode = "pos",
  organism = "Swertia chirayita",
  fourier_components = 0.01,
  frequency = 1,
  resample = 1,
  shift = 0.05,
  time_min = 0.5,
  time_max = 32.5,
  intensity_offset = 100,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection"
)
}
\arguments{
\item{annotations}{Annotations}

\item{features_informed}{Features informed}

\item{features_not_informed}{Features not informed}

\item{file}{File}

\item{headers}{Headers}

\item{detector}{Detector}

\item{show_example}{Show example? Default to FALSE}

\item{min_confidence}{Min confidence}

\item{min_similarity_prefilter}{Min similarity pre filter}

\item{min_similarity_filter}{Min similarity filter}

\item{mode}{Mode}

\item{organism}{Organism}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{resample}{Resample}

\item{shift}{Shift}

\item{time_min}{Time min}

\item{time_max}{Time max}

\item{intensity_offset}{Offset to add to intensity values to handle negative
intensities. Default is 100.}

\item{intensity_floor}{Small value subtracted from minimum intensity.
Default is 0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection".}
}
\value{
A list of plots
}
\description{
Generate pseudochromatograms
}
\examples{
\dontrun{
generate_pseudochromatograms(show_example = TRUE)
}
}
