% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Generate IDs}
\usage{
generate_tables(
  annotations = NULL,
  file_negative = NULL,
  file_positive = NULL,
  min_confidence = 0.4,
  show_example = FALSE,
  export_csv = TRUE,
  export_html = TRUE,
  export_dir = "data/processed",
  export_name = "cascade_table"
)
}
\arguments{
\item{annotations}{Annotations}

\item{file_negative}{File negative}

\item{file_positive}{File positive}

\item{min_confidence}{Min confidence}

\item{show_example}{Show example? Default to FALSE}

\item{export_csv}{Export CSV}

\item{export_html}{Export HTML}

\item{export_dir}{Export Dir}

\item{export_name}{Export name}
}
\value{
Tables
}
\description{
Generate IDs
}
\examples{
\dontrun{
generate_tables()
}
}
