% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msd_meta.R
\name{get_file_metadata}
\alias{get_file_metadata}
\title{Stores all metadata objects to Global Environment}
\usage{
get_file_metadata(file_path)
}
\arguments{
\item{file_path}{path where msd lives}
}
\value{
global objects
}
\description{
Stores all metadata objects to Global Environment
}
\examples{
\dontrun{
merdata <- file.path(glamr::si_path("path_msd"))
file_path <- return_latest(folderpath = merdata, pattern = "Zambia-Daily-2022-08-15")
get_file_metadata(file_path)}

}
