% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{get_pattern_interpretation}
\alias{get_pattern_interpretation}
\title{Get QTL Pattern Interpretation}
\usage{
get_pattern_interpretation(pattern_num)
}
\arguments{
\item{pattern_num}{Numeric pattern number (1-25)}
}
\value{
Character string with the pattern interpretation
}
\description{
Get QTL Pattern Interpretation
}
