% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Get peaks}
\source{
https://github.com/ethanbass/chromatographR
}
\usage{
get_peaks(
  chrom_list,
  lambdas,
  fit = c("egh", "gaussian", "raw"),
  sd.max = 50,
  max.iter = 100,
  time.units = c("min", "s", "ms"),
  estimate_purity = FALSE,
  noise_threshold = 0.001,
  collapse = FALSE,
  ...
)
}
\arguments{
\item{chrom_list}{Chrom list}

\item{lambdas}{Lambdas}

\item{fit}{Fit}

\item{sd.max}{Sd max}

\item{max.iter}{Max iter}

\item{time.units}{Time units}

\item{estimate_purity}{Estimate purity}

\item{noise_threshold}{Noise Threshold}

\item{collapse}{Collapse}

\item{...}{...}
}
\value{
Peaks
}
\description{
Get peaks
}
\note{
This was imported from {chromatographR} package and parallelization
was removed as it was causing issues on Windows.
}
\examples{
NULL
}
\author{
Ethan Bass
}
