% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_subgroup_level_columns}
\alias{get_subgroup_level_columns}
\title{Get Subgroup Level Column Names}
\usage{
get_subgroup_level_columns(hierarchy)
}
\arguments{
\item{hierarchy}{A CellTypeHierarchy object}
}
\value{
List of lists of character vectors
}
\description{
Returns full column names for each subgroup level.
}
\keyword{internal}
