% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchies_grouped_progress.R
\name{hierarchies_grouped_progress}
\alias{hierarchies_grouped_progress}
\title{Title}
\usage{
hierarchies_grouped_progress(xs)
}
\arguments{
\item{xs}{}
}
\description{
Title
}
\examples{
dirty workaround
}
