% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchies_progress.R
\name{hierarchies_progress}
\alias{hierarchies_progress}
\title{Hierarchies Progress}
\usage{
hierarchies_progress(xs, comparison)
}
\arguments{
\item{xs}{XS}

\item{comparison}{Comparison}
}
\value{
A list of hierarchies
}
\description{
Hierarchies Progress
}
\examples{
NULL
}
