% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms_progress.R
\name{histograms_progress}
\alias{histograms_progress}
\title{Histograms progress}
\usage{
histograms_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of histograms
}
\description{
Histograms progress
}
\examples{
NULL
}
