% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_signal.R
\name{improve_signal}
\alias{improve_signal}
\title{Title}
\usage{
improve_signal(
  df,
  fourier_components = FOURRIER_COMPONENTS,
  time_min = TIME_MIN,
  time_max = TIME_MAX,
  frequency = FREQUENCY,
  resample = RESAMPLE
)
}
\arguments{
\item{resample}{}
}
\value{

}
\description{
Title
}
