% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{initial}
\alias{initial}
\title{Initial values for model}
\usage{
initial(p, ...)
}
\arguments{
\item{p}{parameters (from \code{\link{parameters}})}

\item{...}{vector of named arguments}
}
\description{
Initial values for model.
}

