% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{INTENSITY_MS_MIN}
\alias{INTENSITY_MS_MIN}
\title{Parameters related to MS/CAD}
\format{
An object of class \code{integer} of length 1.
}
\usage{
INTENSITY_MS_MIN
}
\description{
Parameters related to MS/CAD
}
\keyword{datasets}
