% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_peaks.R
\name{join_peaks}
\alias{join_peaks}
\title{Join peaks}
\usage{
join_peaks(chromatograms, peaks, min_area)
}
\arguments{
\item{chromatograms}{Chromatograms}

\item{peaks}{Peaks}

\item{min_area}{Min area}
}
\value{
A dataframe with joined peaks
}
\description{
Join peaks
}
\examples{
NULL
}
