% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_munging.R
\name{lag_tx_curr}
\alias{lag_tx_curr}
\title{Create TX_CURR_Lag2
Filters the cascade variables to current fiscal year}
\usage{
lag_tx_curr(df)
}
\arguments{
\item{.data}{MER structured data set}
}
\value{
passed through a filtered cascade data frame
}
\description{
Create TX_CURR_Lag2
Filters the cascade variables to current fiscal year
}
