% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{LFSR_NULL_THRESHOLD}
\alias{LFSR_NULL_THRESHOLD}
\title{LFSR Null Hypothesis Threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LFSR_NULL_THRESHOLD
}
\description{
LFSR Null Hypothesis Threshold
}
\keyword{datasets}
