% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{LFSR_SIG_THRESHOLD}
\alias{LFSR_SIG_THRESHOLD}
\title{LFSR Significance Threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LFSR_SIG_THRESHOLD
}
\description{
LFSR Significance Threshold
}
\keyword{datasets}
