% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{load_annotations}
\alias{load_annotations}
\title{Load annotations}
\usage{
load_annotations(file = NULL, show_example = FALSE, mode = "pos")
}
\arguments{
\item{file}{File}

\item{show_example}{Show example? Default to FALSE}

\item{mode}{Mode}
}
\value{
A table of annotations
}
\description{
Load annotations
}
\examples{
NULL
}
