% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_features.R
\name{load_features}
\alias{load_features}
\title{Load features}
\usage{
load_features(file = NULL, show_example = FALSE)
}
\arguments{
\item{file}{File}

\item{show_example}{Show example? Default to FALSE}
}
\value{
A table of features
}
\description{
Load features
}
\examples{
NULL
}
