% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_features_not_informed.R
\name{load_features_not_informed}
\alias{load_features_not_informed}
\title{Load features not informed}
\usage{
load_features_not_informed(file = NULL, show_example = FALSE)
}
\arguments{
\item{file}{File}

\item{show_example}{Show example? Default to FALSE}
}
\value{
A table of non informed features
}
\description{
Load features not informed
}
\examples{
NULL
}
