% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ms_data.R
\name{load_ms_data}
\alias{load_ms_data}
\title{Load MS data}
\usage{
load_ms_data(file = NULL, show_example = FALSE)
}
\arguments{
\item{file}{File}

\item{show_example}{Show example? Default to FALSE}
}
\value{
MS data
}
\description{
Load MS data
}
\examples{
NULL
}
