% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_name.R
\name{load_name}
\alias{load_name}
\title{Load name}
\usage{
load_name(
  file = NULL,
  default = "210619_AR_06_V_03_2_01.mzML",
  show_example = FALSE
)
}
\arguments{
\item{file}{File}

\item{default}{Default}

\item{show_example}{Show example? Default to FALSE}
}
\value{
A name
}
\description{
Load name
}
\examples{
NULL
}
