% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chromatographiable.R
\name{make_chromatographiable}
\alias{make_chromatographiable}
\title{Title}
\usage{
make_chromatographiable(
  df,
  mass_min = params$structures$mass$min,
  mass_max = params$structures$mass$max,
  logp_min = params$structures$logp$min,
  logp_max = params$structures$logp$max
)
}
\arguments{
\item{logp_max}{}
}
\description{
Title
}
