% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confident.R
\name{make_confident}
\alias{make_confident}
\title{Make confident}
\usage{
make_confident(df, score)
}
\arguments{
\item{df}{Dataframe}

\item{score}{Score}
}
\value{
A dataframe containing annotations with scores above the confidence threshold set
}
\description{
Make confident
}
\examples{
NULL
}
