% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_no_stereo.R
\name{make_no_stereo}
\alias{make_no_stereo}
\title{Make no stereo}
\usage{
make_no_stereo(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
A dataframe with no stereo structures
}
\description{
Make no stereo
}
\examples{
NULL
}
