% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_other.R
\name{make_other}
\alias{make_other}
\title{Make other}
\usage{
make_other(dataframe, value = "peak_area")
}
\arguments{
\item{dataframe}{Dataframe}

\item{value}{Value}
}
\value{
A dataframe with harmonized "other" subcategories
}
\description{
Make other
}
\examples{
NULL
}
