% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade-package.R
\docType{class}
\name{micro_array-class}
\alias{micro_array-class}
\title{Class \code{"micro_array"}}
\description{
The Class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("micro_array", ...)}. %% ~~ describe objects here ~~
}

\examples{

showClass("micro_array")

}
\keyword{classes}
