\name{micropredict-class}
\Rdversion{1.1}
\docType{class}
\alias{micropredict-class}


\title{Class \code{"micropredict"}}
\description{
Class for prediction of microarray value.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("micropredict", ...)}.
}

\examples{
showClass("micropredict")
}
\keyword{classes}
