% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade-package.R
\docType{class}
\name{micropredict-class}
\alias{micropredict-class}
\title{Class \code{"micropredict"}}
\description{
Class for prediction of microarray value.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("micropredict", ...)}.
}

\examples{

showClass("micropredict")

}
\keyword{classes}
