% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Net_inf}
\alias{Net_inf}
\title{Reverse-engineered network of the simulated data.}
\description{
The reverse-engineered network of the simulated data (M and Net).
}
\examples{

data(Net_inf)
str(Net_inf)

}
\keyword{datasets}
