% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade-package.R
\docType{class}
\name{network-class}
\alias{network-class}
\title{Class \code{"network"}}
\description{
The \code{"network"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("network", ...)}. %% ~~ describe objects here ~~
}

\examples{

showClass("network")

}
\keyword{classes}
