% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{network}
\alias{network}
\title{A network object data.}
\description{
A network object. It is the same as the result in the vignette for the
inference of the network.
}
\examples{

data(network)
plot(network)
print(network)

}
\keyword{datasets}
