% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_chromato.R
\name{normalize_chromato}
\alias{normalize_chromato}
\title{Normalize chromato}
\usage{
normalize_chromato(x, list = list)
}
\arguments{
\item{x}{X}

\item{list}{List}
}
\value{
A normalized chromato
}
\description{
Normalize chromato
}
\examples{
NULL
}
