% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_chromatograms_list.R
\name{normalize_chromatograms_list}
\alias{normalize_chromatograms_list}
\title{Title}
\usage{
normalize_chromatograms_list(list, shift = 0, time = FALSE, intensity = TRUE)
}
\arguments{
\item{intensity}{}
}
\description{
Title
}
