% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_acn.R
\name{p_acn_i}
\alias{p_acn_i}
\title{P ACN I}
\usage{
p_acn_i(acn_eluent, q1, q2, q3)
}
\arguments{
\item{acn_eluent}{ACN eluent}

\item{q1}{Q1}

\item{q2}{Q2}

\item{q3}{Q3}
}
\value{
P ACN I
}
\description{
P ACN I
}
\examples{
NULL
}
