% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_progress.R
\name{peaks_progress}
\alias{peaks_progress}
\title{Peaks progress}
\usage{
peaks_progress(df_xy)
}
\arguments{
\item{df_xy}{Df X Y}
}
\value{
A list of peaks
}
\description{
Peaks progress
}
\examples{
NULL
}
