% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromatogram.R
\name{plot_chromatogram}
\alias{plot_chromatogram}
\title{Plot chromatogram}
\usage{
plot_chromatogram(df, text)
}
\arguments{
\item{df}{Dataframe}

\item{text}{Text}
}
\value{
A plot of a chromatogram
}
\description{
Plot chromatogram
}
\examples{
NULL
}
