% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Plot histograms}
\usage{
plot_histograms(dataframe, label, y = "values", xlab = TRUE)
}
\arguments{
\item{dataframe}{Dataframe}

\item{label}{Label}

\item{y}{Y}

\item{xlab}{Xlab}
}
\value{
A plot of histograms
}
\description{
Plot histograms
}
\examples{
NULL
}
