% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms_litt}
\alias{plot_histograms_litt}
\title{Plot histograms litt}
\usage{
plot_histograms_litt(dataframe, label, y = "values", xlab = TRUE)
}
\arguments{
\item{dataframe}{Dataframe}

\item{label}{Label}

\item{y}{Y}

\item{xlab}{Xlab}
}
\value{
A plot of literature histograms
}
\description{
Plot histograms litt
}
\examples{
NULL
}
