% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms_taxo}
\alias{plot_histograms_taxo}
\title{Plot histograms taxo}
\usage{
plot_histograms_taxo(
  dataframe,
  chromatogram,
  level = "max",
  mode = "pos",
  time_min,
  time_max
)
}
\arguments{
\item{dataframe}{Dataframe}

\item{chromatogram}{Chromatogram}

\item{level}{Level}

\item{mode}{Mode}

\item{time_min}{Time min}

\item{time_max}{Time max}
}
\value{
A plot of taxo histograms
}
\description{
Plot histograms taxo
}
\examples{
NULL
}
