% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak_detection.R
\name{plot_peak_detection}
\alias{plot_peak_detection}
\title{Plot peak detection}
\usage{
plot_peak_detection(df1, df2, fun)
}
\arguments{
\item{df1}{DF 1 containing chromatogram}

\item{df2}{DF 2 containing peaks}

\item{fun}{Fun}
}
\value{
A plot with (non-)detected peaks
}
\description{
Plot peak detection
}
\examples{
NULL
}
