% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peaks_statistics.R
\name{plot_peaks_statistics}
\alias{plot_peaks_statistics}
\title{Plot peaks statistics}
\usage{
plot_peaks_statistics(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
A plot with peaks statistics
}
\description{
Plot peaks statistics
}
\examples{
NULL
}
