% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results_2}
\alias{plot_results_2}
\title{Plot results 2}
\usage{
plot_results_2(detector = "cad")
}
\arguments{
\item{detector}{Detector}
}
\value{
A list of plots
}
\description{
Plot results 2
}
\examples{
NULL
}
