% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tima_histograms.R
\name{plot_tima_histograms}
\alias{plot_tima_histograms}
\title{Plot TIMA histograms}
\usage{
plot_tima_histograms(tables)
}
\arguments{
\item{tables}{Tables}
}
\value{
Pretty plots
}
\description{
Plot TIMA histograms
}
\examples{
NULL
}
