% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\docType{methods}
\name{position-methods}
\alias{position-methods}
\alias{position}
\alias{position,network-method}
\title{Returns the position of edges in the network}
\usage{
\S4method{position}{network}(net, nv = 0)
}
\arguments{
\item{net}{a network object}

\item{nv}{the level of cutoff at which the analysis should be done}
}
\description{
Returns the position of edges in the network
}
\section{Methods}{
 \describe{

\item{list("signature(net = \"network\")")}{ Returns a matrix with the
position of the node. This matrix can then be used as an argument in the
plot function. } }
}

\examples{

data(Net)
position(Net)

}
\keyword{methods}
