% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prehistograms_progress.R
\name{prehistograms_progress}
\alias{prehistograms_progress}
\title{Prehistograms progress}
\usage{
prehistograms_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of prehistograms
}
\description{
Prehistograms progress
}
\examples{
NULL
}
