% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comparison.R
\name{prepare_comparison}
\alias{prepare_comparison}
\title{Prepare comparison}
\usage{
prepare_comparison(
  features_informed = NULL,
  features_not_informed = NULL,
  candidates_confident,
  min_similarity_prefilter = 0.6,
  min_similarity_filter = 0.8,
  mode = "pos",
  show_example = FALSE,
  default_peak_area = 0.001
)
}
\arguments{
\item{features_informed}{Features informed}

\item{features_not_informed}{Features not informed}

\item{candidates_confident}{Candidates confident}

\item{min_similarity_prefilter}{Min similarity pre filter}

\item{min_similarity_filter}{Min similarity filter}

\item{mode}{Mode}

\item{show_example}{Show example? Default to FALSE}

\item{default_peak_area}{Default peak area for features without peak
information. Default is 0.001.}
}
\value{
A list of peaks
}
\description{
Prepare comparison
}
\examples{
NULL
}
