% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features.R
\name{prepare_features}
\alias{prepare_features}
\title{Prepare features}
\usage{
prepare_features(df, min_intensity)
}
\arguments{
\item{df}{}
}
\value{
A dataframe of prepared features
}
\description{
Prepare features
}
\examples{
NULL
}
