% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy.R
\name{prepare_hierarchy}
\alias{prepare_hierarchy}
\title{Prepare hierarchy}
\usage{
prepare_hierarchy(
  dataframe,
  type = "analysis",
  detector = "ms",
  rescale = FALSE
)
}
\arguments{
\item{dataframe}{Dataframe}

\item{type}{Type}

\item{detector}{Detector}

\item{rescale}{Rescale}
}
\value{
A dataframe with prepared hierarchy
}
\description{
Prepare hierarchy
}
\examples{
NULL
}
