% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_preparation.R
\name{prepare_hierarchy_preparation}
\alias{prepare_hierarchy_preparation}
\title{Title}
\usage{
prepare_hierarchy_preparation(dataframe)
}
\arguments{
\item{dataframe}{}
}
\description{
Title
}
