% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot.R
\name{prepare_plot}
\alias{prepare_plot}
\title{Prepare plot}
\usage{
prepare_plot(dataframe, organism = "species")
}
\arguments{
\item{dataframe}{Dataframe}

\item{organism}{Organism}
}
\value{
A dataframe prepared for plots
}
\description{
Prepare plot
}
\examples{
NULL
}
