% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot.R
\name{prepare_plot_2}
\alias{prepare_plot_2}
\title{Prepare plot 2}
\usage{
prepare_plot_2(dataframe)
}
\arguments{
\item{dataframe}{Dataframe}
}
\value{
A dataframe prepared for plots
}
\description{
Prepare plot 2
}
\examples{
NULL
}
