% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tima_annotations.R
\name{prepare_tima_annotations}
\alias{prepare_tima_annotations}
\title{Prepare TIMA annotations}
\usage{
prepare_tima_annotations(
  annotations = NULL,
  predicted_classes = FALSE,
  min_score_initial = 0,
  min_score_biological = 0,
  min_score_chemical = 0,
  min_score_final = 0,
  min_matched_peaks_absolute = 0L,
  min_matched_peaks_percentage = 0,
  min_peaks = 3L,
  libraries = c("gnps", "massbank", "merlin", "ISDB", "ISDB - Wikidata", "TIMA MS1"),
  show_example = FALSE
)
}
\arguments{
\item{annotations}{annotations}

\item{predicted_classes}{Show predicted classes? Default to FALSE}

\item{min_score_initial}{Minimal initial score}

\item{min_score_biological}{Minimal biological score}

\item{min_score_chemical}{Minimal chemical score}

\item{min_score_final}{Minimal final score}

\item{min_matched_peaks_absolute}{Minimal number of matched peaks}

\item{min_matched_peaks_percentage}{Minimal percentage of matched peaks}

\item{min_peaks}{Minimal number of peaks in spectrum}

\item{libraries}{Libraries to consider}

\item{show_example}{Show example? Default to FALSE}
}
\value{
Prepared tables
}
\description{
Prepare TIMA annotations
}
\examples{
NULL
}
