% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyTables_progress.R
\name{prettyTables_progress}
\alias{prettyTables_progress}
\title{Pretty tables progress}
\usage{
prettyTables_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
Pretty tables
}
\description{
Pretty tables progress
}
\examples{
NULL
}
