% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R, R/network.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,micro_array-method}
\alias{print,network-method}
\title{Methods for Function \code{print}}
\usage{
\S4method{print}{micro_array}(x, ...)

\S4method{print}{network}(x, ...)
}
\arguments{
\item{x}{an object of class micro-array or network}

\item{...}{additional parameters}
}
\description{
Methods for function \code{print}
}
\examples{

data(Net)
print(Net)

data(M)
print(M)

}
\keyword{methods}
