% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{print.CellTypeHierarchy}
\alias{print.CellTypeHierarchy}
\title{Print method for CellTypeHierarchy}
\usage{
\method{print}{CellTypeHierarchy}(x, ...)
}
\arguments{
\item{x}{A CellTypeHierarchy object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for CellTypeHierarchy
}
