% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_compare_peaks.R
\name{process_compare_peaks}
\alias{process_compare_peaks}
\title{Process compare peaks}
\usage{
process_compare_peaks(
  file = NULL,
  features = NULL,
  chromatogram = "baselined",
  detector = "cad",
  export_dir = "data/interim/peaks",
  show_example = FALSE,
  fourier_components = 0.01,
  frequency = 2,
  min_area = 0.005,
  min_intensity = 10000,
  resample = 1,
  shift = 0.05,
  time_min = 0,
  time_max = Inf
)
}
\arguments{
\item{file}{File path}

\item{show_example}{Show example? Default to FALSE}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{resample}{Resample}

\item{time_min}{Minimum time}

\item{time_max}{Maximum time}

\item{file_positive}{Positive file path}

\item{cad_shift}{CAD shift}

\item{pda_shift}{PDA shift}

\item{chromatograms}{Chromatograms to plot}

\item{type}{Type. "baselined" or "improved"}

\item{normalize_intensity}{Normalize intensity? Default to TRUE}

\item{normalize_time}{Normalize time? Default to FALSE}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Process compare peaks
}
\examples{
\dontrun{
check_chromatograms_alignment(show_example = TRUE)
}
}
