% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peaks.R
\name{process_peaks}
\alias{process_peaks}
\title{Process peaks}
\usage{
process_peaks(detector = "cad")
}
\arguments{
\item{detector}{}
}
\value{
A list of processed peaks
}
\description{
Process peaks
}
\examples{
NULL
}
