% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plot_pseudochromatograms.R
\name{process_plot_pseudochromatograms}
\alias{process_plot_pseudochromatograms}
\title{Process plot pseudochromatograms}
\usage{
process_plot_pseudochromatograms()
}
\value{
A plot with histograms on chromatograms
}
\description{
Process plot pseudochromatograms
}
\examples{
\dontrun{
process_plot_pseudochromatograms(show_example = TRUE)
}
export
}
