% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{QTL_MECHANISMS}
\alias{QTL_MECHANISMS}
\title{Scientific Definitions}
\format{
An object of class \code{character} of length 8.
}
\usage{
QTL_MECHANISMS
}
\description{
QTL mechanism categories, pattern definitions, variant heterogeneity
  codes, and other scientific constants used in categorization.
QTL Mechanism Categories

Eight main categories for variant QTL mechanisms
}
\keyword{datasets}
