% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{QTL_PATTERNS}
\alias{QTL_PATTERNS}
\title{QTL Pattern Details}
\format{
An object of class \code{list} of length 25.
}
\usage{
QTL_PATTERNS
}
\description{
Maps 25 QTL patterns to their interpretations and mechanism categories
Each pattern has: interpretation (detailed description) and mechanism (category index)
}
\keyword{datasets}
