% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_cascade.R
\name{return_cascade}
\alias{return_cascade}
\title{Returns a munged data frame of cascade data.}
\usage{
return_cascade(df, cscd_num)
}
\arguments{
\item{df}{MER structured data set for a given population and geography}

\item{cscd_num}{cascade selection from user}
}
\value{
a data frame of cascade indicators
}
\description{
Fetches the correct version of the data from each cascade option.
Returns a data frame of cascade indicators for a given population
}
\examples{
\dontrun{
# returns standard cascade
return_cascade(df, 1)

# returns KP cascade
return_cascade(df, 13)
}

}
