% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{save_categorization_summary}
\alias{save_categorization_summary}
\title{Save categorization summary}
\usage{
save_categorization_summary(summary_data, output_file)
}
\arguments{
\item{summary_data}{Summary data frame}

\item{output_file}{Output file path}
}
\description{
Save categorization summary
}
\keyword{internal}
