% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_histograms_progress.R
\name{save_histograms_progress}
\alias{save_histograms_progress}
\title{Save histograms progress}
\usage{
save_histograms_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
Saved histograms
}
\description{
Save histograms progress
}
\examples{
NULL
}
