% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_prettySubtables_progress.R
\name{save_prettySubtables_progress}
\alias{save_prettySubtables_progress}
\title{Save pretty subtables progress}
\usage{
save_prettySubtables_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
Saved pretty subtables
}
\description{
Save pretty subtables progress
}
\examples{
NULL
}
