% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_treemaps_progress.R
\name{save_treemaps_progress}
\alias{save_treemaps_progress}
\title{Save treemaps progress}
\usage{
save_treemaps_progress(xs, type = "treemap")
}
\arguments{
\item{xs}{XS}

\item{type}{Type}
}
\value{
Saved treemaps
}
\description{
Save treemaps progress
}
\examples{
NULL
}
