% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCADE.R
\name{sc_CASCADE}
\alias{sc_CASCADE}
\title{CASCADE calculation for single cell samples.}
\usage{
sc_CASCADE(
  total_obj = NULL,
  tumor_obj = NULL,
  nontumor_obj = NULL,
  malignancy_col = "malignancy",
  ident_col = "ident.name",
  subtype_col = "subtype",
  min_n_cells = 15,
  subtype_scores = eco_subtype_scores,
  lineage_genes = top_lineage_genes
)
}
\arguments{
\item{total_obj}{Seurat object containing all cells. Either this parameter
or both \code{tumor_obj} and \code{nontumor_obj} must be supplied. If \code{total_obj} is
given, then this object must contain a metadata column with malignancy info
specified by the \code{malignancy_col} parameter.}

\item{tumor_obj}{Seurat object containing all malignant cells. Ignored if
\code{total_obj} is supplied. If \code{total_obj} is not supplied, then both
\code{tumor_obj} and \code{nontumor_obj} must be supplied.}

\item{nontumor_obj}{Seurat object containing all non-malignant cells.
Ignored if \code{total_obj} is supplied. If \code{total_obj} is not supplied, then
both \code{tumor_obj} and \code{nontumor_obj} must be supplied.}

\item{malignancy_col}{Column name for metadata column containing malignancy
information for \code{total_obj}. This column must be a string vector with
"malignant" and "non-malignant" labels. Not needed if \code{total_obj} is null.}

\item{ident_col}{Name of the metadata column containing a sample label for
each cell. Must be included in the metadata of each supplied Seurat object.}

\item{subtype_col}{Name of the metadata column containing a subtype label for
each cell. Must be included in the metadata of \code{total_obj} or \code{nontumor_obj}.}

\item{min_n_cells}{Integer number of the minimum required cell count for
each sample. Samples with fewer than this number of cells will be excluded.}

\item{subtype_scores}{Data frame containing nontumor cell subtype weights
and labels. Rownames are subtypes. Must contain a \code{category} column that
labels each subtype as a "Hazard" or "Resource" and a \code{score} column that
lists a weight for each subtype. Defaults to a built-in weight table curated
from a single-cell study of liver cancer.}

\item{lineage_genes}{List of vectors containing genes for each lineage.
Gene set vectors must be named after the lineages that they represent.
Defaults to a built-in list of genes curated from a single-cell study of
liver cancer. Contains gene sets for CLTCs (cholangiocyte-like tumor cells),
HLTCs (hepatocyte-like tumor cells), MCTCs (myeloid cell-like tumor cells),
and MLTCs (mesenchymal-like tumor cells). If an independent value is given
for this parameter, it must contain all 4 of these lineages in order to
calculate a lineage score.}
}
\value{
Data frame with the following columns:
CLTC: expression of cholangiocyte-like tumor cell markers in tumor cells
HLTC: expression of hepatocyte-like tumor cell markers in tumor cells
MCTC: expression of myeloid cell-like tumor cell markers in tumor cells
MLTC: expression of mesenchymal-like tumor cell markers in tumor cells
additional columns for any other lineages supplied in \code{lineage_genes}
lineage_score: lineage score in each sample
hazards: amount of hazards facing the non-malignant cells
resources: amount of resources available to the non-malignant cells
eco_score: eco score in each sample
tumor_score: tumor score in each sample
quadrant: lineage-eco quadrant for each sample
}
\description{
CASCADE calculation for single cell samples.
}
\examples{
sc_CASCADE(total_obj = x, malignancy_col = "mal_anno", ident_col = "sample")
sc_CASCADE(tumor_obj = t, nontumor_obj = n, subtype_col = "cell_type")

}
