% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCADE.R
\name{sc_LINEAGE}
\alias{sc_LINEAGE}
\title{Lineage score calculation for single cell samples}
\usage{
sc_LINEAGE(expr_data, idents, lineage_genes = top_lineage_genes)
}
\arguments{
\item{expr_data}{Matrix of normalized gene counts from a single-cell dataset.
Genes are on the rows and samples are on the columns.}

\item{idents}{Vector of sample labels for the columns of \code{expr_data}.}

\item{lineage_genes}{List of vectors containing genes for each lineage.
Gene set vectors must be named after the lineages that they represent.
Defaults to a built-in list of genes curated from a single-cell study of
liver cancer. Contains gene sets for CLTCs (cholangiocyte-like tumor cells),
HLTCs (hepatocyte-like tumor cells), MCTCs (myeloid cell-like tumor cells),
and MLTCs (mesenchymal-like tumor cells). If an independent value is given
for this parameter, it must contain all 4 of these lineages in order to
calculate a lineage score.}
}
\value{
Data frame containing average gene expression for each lineage and
the final lineage score in the "lineage_score" column. Samples are row names.
}
\description{
Lineage score reflects characteristics of a tumor cell population. It is
uniformly distributed between 0 and 1, where a low score indicates a
hepatocytic or myeloid cell-like lineage and a high score indicates a
cholangiocytic or mesenchymal lineage.
}
\examples{
expr_data: cell1 cell2 cell3 ...
     gene1     0   0.1   1.2
     gene2   0.9     0     0
     gene3   2.2   2.6   1.5
     .
     .
     .
idents <- c("Sample1", "Sample1", "Sample2")
lineage_genes <- list(CLTC = c("gene12", "gene23", "gene45"),
                      HLTC = c("gene1", "gene91", "gene92", "gene93"),
                      MCTC = c("gene2", "gene96", "gene97", "gene98"),
                      MLTC = c("gene5", "gene6", "gene14"))

sc_LINEAGE(expr_data, idents, lineage_genes)

}
