% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{sigma}
\alias{sigma}
\title{Parameters adapted from Excel sheet from paper shortDOI: 10/ghmvhz}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sigma
}
\description{
Parameters adapted from Excel sheet from paper shortDOI: 10/ghmvhz
}
\keyword{datasets}
