% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sharpening.R
\name{signal_sharpening}
\alias{signal_sharpening}
\title{Title}
\usage{
signal_sharpening(
  Time = timeow,
  Intensity = intensityeah,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  Smoothing_width = 8,
  Baseline_adjust = 0
)
}
\arguments{
\item{Baseline_adjust}{}
}
\description{
Title
}
