\name{simul}
\alias{simul}
\alias{M}
\alias{Net}
\alias{Net_inf}
\docType{data}
\title{
Simulated data for examples.
}
\description{
Three objects: M, microarray, Net, network, and Net_inf, the reverse-engineered network.
}
\usage{data(simul)}

\examples{
data(simul)
head(M)
str(Net)
str(Net_inf)
}
\keyword{datasets}