% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtables_progress.R
\name{subtables_progress}
\alias{subtables_progress}
\title{Subtables progress}
\usage{
subtables_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of subtables
}
\description{
Subtables progress
}
\examples{
NULL
}
