% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,micro_array-method}
\title{Methods for Function \code{summary}}
\usage{
\S4method{summary}{micro_array}(object, nb.graph = NULL, ...)
}
\arguments{
\item{object}{an object of class micro-array}

\item{nb.graph}{(optionnal) choose the graph to plot. Displays all graphs by default.}

\item{...}{additional parameters.}
}
\description{
Methods for function \code{summary}
}
\examples{

data(M)
summary(M)

}
\keyword{methods}
