% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name_to_qid.R
\name{taxon_name_to_qid}
\alias{taxon_name_to_qid}
\title{Taxon name to QID}
\usage{
taxon_name_to_qid(taxon_name)
}
\arguments{
\item{taxon_name}{Taxon name}
}
\value{
A QID
}
\description{
Taxon name to QID
}
\examples{
\dontrun{
taxon_name_to_qid(taxon_name = "Gentiana lutea")
}
}
