% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt.R
\name{temp_gt_function}
\alias{temp_gt_function}
\title{Temp GT function}
\usage{
temp_gt_function(table, title, subtitle)
}
\arguments{
\item{table}{Table}

\item{title}{Title}

\item{subtitle}{Subtitle}
}
\value{
A formatted GT table
}
\description{
Temp GT function
}
\examples{
NULL
}
