% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{TIME_MIN}
\alias{TIME_MIN}
\title{Parameters for LC alignment}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TIME_MIN
}
\description{
Parameters for LC alignment
}
\keyword{datasets}
