% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_baseline.R
\name{transform_baseline}
\alias{transform_baseline}
\title{Transform baseline}
\usage{
transform_baseline(x)
}
\arguments{
\item{x}{X}
}
\value{
A list with transformed baseline
}
\description{
Transform baseline
}
\examples{
NULL
}
