% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_ms.R
\name{transform_ms}
\alias{transform_ms}
\title{Transform MS}
\usage{
transform_ms(x, min_intensity = 0.1)
}
\arguments{
\item{x}{X}

\item{min_intensity}{Minimum normalized intensity threshold for filtering.
Default is 0.1. Set to 0 to keep all points.}
}
\value{
A list with transformed MS
}
\description{
Transform MS
}
\examples{
NULL
}
