% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemaps_progress.R
\name{treemaps_progress}
\alias{treemaps_progress}
\title{Treemaps progress}
\usage{
treemaps_progress(xs, type = "treemap", hierarchies)
}
\arguments{
\item{xs}{XS}

\item{type}{Type}

\item{hierarchies}{Hierarchies}
}
\value{
A list of treemaps
}
\description{
Treemaps progress
}
\examples{
NULL
}
