% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemaps_progress.R
\name{treemaps_progress_no_title}
\alias{treemaps_progress_no_title}
\title{Treemaps progress no title}
\usage{
treemaps_progress_no_title(xs, type = "treemap")
}
\arguments{
\item{xs}{XS}

\item{type}{Type}
}
\value{
A list of treemaps with no title
}
\description{
Treemaps progress no title
}
\examples{
NULL
}
