% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{VARIANT_HETEROGENEITY}
\alias{VARIANT_HETEROGENEITY}
\title{Variant Heterogeneity Categories}
\format{
An object of class \code{character} of length 4.
}
\usage{
VARIANT_HETEROGENEITY
}
\description{
Categories for variant heterogeneity in multi-cell-type features.
Maps category names to letter codes (a-d)
}
\keyword{datasets}
