% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_progress.R
\name{wiki_progress}
\alias{wiki_progress}
\title{Wiki progress}
\usage{
wiki_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of results of Wikidata queries
}
\description{
Wiki progress
}
\examples{
NULL
}
