% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_paths_params.R
\docType{data}
\name{WORKERS}
\alias{WORKERS}
\title{Generic parameters}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WORKERS
}
\description{
Generic parameters
}
\keyword{datasets}
