% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chromato_line.R
\name{add_chromato_line}
\alias{add_chromato_line}
\title{Title}
\usage{
add_chromato_line(
  plot,
  chromato,
  shift = 0,
  time,
  name,
  color,
  polarity = "pos"
)
}
\arguments{
\item{polarity}{}
}
\description{
Title
}
