% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peak_metadata.R
\name{add_peak_metadata}
\alias{add_peak_metadata}
\title{Add peak metadata}
\usage{
add_peak_metadata(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
A dataframe with peak metadata
}
\description{
Add peak metadata
}
\examples{
NULL
}
