% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_chromatogram.R
\name{baseline_chromatogram}
\alias{baseline_chromatogram}
\title{Baseline chromatogram}
\usage{
baseline_chromatogram(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
A dataframe with baselined chromatogram
}
\description{
Baseline chromatogram
}
\examples{
NULL
}
