% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCADE.R
\name{bulk_CASCADE}
\alias{bulk_CASCADE}
\title{CASCADE calculation for bulk transcriptomic samples.}
\usage{
bulk_CASCADE(bulk_expr, degs = cascade_degs)
}
\arguments{
\item{bulk_expr}{Matrix of normalized gene counts from bulk dataset.}

\item{degs}{Built-in gene set to calculate CASCADE values in bulk datasets.}
}
\description{
CASCADE calculation for bulk transcriptomic samples.
}
