% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade-package.R
\docType{package}
\name{cascade-package}
\alias{cascade}
\alias{cascade-package}
\title{cascade: Contextualizing untargeted Annotation with Semi-quantitative Charged Aerosol Detection for pertinent characterization of natural Extracts}
\description{
This package provides the infrastructure to perform Automated Composition Assessment of Natural Extracts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adafede/cascade}
  \item \url{https://adafede.github.io/cascade}
  \item Report bugs at \url{https://github.com/adafede/cascade/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{rutz@imsb.biol.ethz.ch} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

}
\keyword{internal}
