\name{Cascade-package}
\alias{Cascade-package}
\alias{Cascade}
\docType{package}
\title{The Cascade Package}
\description{
A modeling tool allowing gene selection, reverse engineering, and prediction in Cascade networks.
}
\details{
\tabular{ll}{
Package: \tab Cascade\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2019-02-09\cr
License: \tab GNU 2.0\cr
Depends: \tab methods\cr
}

}
\author{
This package has been written by Frédéric Bertrand, Myriam Maumy-Bertrand and Nicolas Jung with biological insights from Laurent Vallat.
Maintainer:  Frédéric Bertrand <frederic.bertrand@math.unistra.fr>
}


\references{
Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2014). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. \emph{Bioinformatics}, btt705. 
	
Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. \emph{Proceedings of the National Academy of Sciences}, 110(2), 459-464.
}

\keyword{ package }

