% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade-package.R
\docType{package}
\name{Cascade-package}
\alias{Cascade-package}
\alias{Cascade}
\title{The Cascade Package: Selection, Reverse-Engineering and Prediction in Cascade Networks}
\description{
A modeling tool allowing gene selection, reverse engineering, and prediction in cascade networks. Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2014) <doi:10.1093/bioinformatics/btt705>.
}
\references{
Jung, N., Bertrand, F., Bahram, S., Vallat, L., and
Maumy-Bertrand, M. (2014). Cascade: a R-package to study, predict and
simulate the diffusion of a signal through a temporal gene network.
\emph{Bioinformatics}, btt705.

Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F.,
Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic
circuitry of a cancer cell with predicted intervention in chronic
lymphocytic leukemia. \emph{Proceedings of the National Academy of
Sciences}, 110(2), 459-464.
}
\author{
This package has been written by Frédéric Bertrand, Myriam
Maumy-Bertrand and Nicolas Jung with biological insights from Laurent
Vallat. Maintainer: Frédéric Bertrand <frederic.bertrand@math.unistra.fr>
}
\keyword{package}
