\name{Cascade-package}
\alias{Cascade-package}
\alias{Cascade}
\docType{package}
\title{The Cascade Package}
\description{
The Cascade is a modeling tool allowing gene selection, reverse engineering, and prediction.
}
\details{
\tabular{ll}{
Package: \tab Cascade\cr
Type: \tab Package\cr
Version: \tab 1.03\cr
Date: \tab 2014-03-24\cr
License: \tab GNU 2.0\cr
Depends: \tab methods\cr
}

}
\author{
This package has been written by Nicolas Jung in collaboration with Frederic Bertrand, Myriam Maumy-Bertrand and Laurent Vallat.
Maintainer:  <cascade@math.unistra.fr>
}


\references{
Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2013). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. \emph{Bioinformatics}, btt705. 
	
Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. \emph{Proceedings of the National Academy of Sciences}, 110(2), 459-464.
}

\keyword{ package }

