% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calib_model.R
\name{cascade_calib_derivs}
\alias{cascade_calib_derivs}
\title{Compute derivatives for cascade model (calibration)}
\usage{
cascade_calib_derivs(y, p, i)
}
\arguments{
\item{y}{initial state variable values.}

\item{p}{list containing parameters.}

\item{i}{list containing incidence.}
}
\description{
Compute derivatives for cascade model (calibration).
}

