% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chromatograms.R
\name{check_chromatograms}
\alias{check_chromatograms}
\title{Title}
\usage{
check_chromatograms(
  chromatograms = c("bpi_pos", "cad_pos", "pda_pos", "bpi_neg"),
  normalize_time = FALSE,
  shift_cad = 0,
  shift_pda = 0,
  type = "improved"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
