% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chromatograms_alignment.R
\name{check_chromatograms_alignment}
\alias{check_chromatograms_alignment}
\title{Check chromatograms alignment}
\usage{
check_chromatograms_alignment(
  file_negative = NULL,
  file_positive = NULL,
  time_min = 0.5,
  time_max = 32.5,
  cad_shift = 0.05,
  pda_shift = 0.1,
  fourier_components = 0.01,
  frequency = 1,
  resample = 1,
  chromatograms = c("bpi_pos", "cad_pos", "pda_pos"),
  headers = c(bpi = "BasePeak_0", pda = "PDA#1_TotalAbsorbance_0", cad = "UV#1_CAD_1_0"),
  type = "baselined",
  normalize_intensity = TRUE,
  normalize_time = FALSE,
  show_example = FALSE,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection",
  improve_signal = TRUE
)
}
\arguments{
\item{file_negative}{Negative file path}

\item{file_positive}{Positive file path}

\item{time_min}{Minimum time in minutes. Default is 0.5.}

\item{time_max}{Maximum time in minutes. Default is 32.5.}

\item{cad_shift}{CAD time shift in minutes. Default is 0.05.}

\item{pda_shift}{PDA time shift in minutes. Default is 0.1.}

\item{fourier_components}{Fraction of Fourier components to keep. Default is
0.01.}

\item{frequency}{Acquisition frequency in Hz. Default is 1.}

\item{resample}{Resampling factor. Default is 1.}

\item{chromatograms}{Chromatograms to plot. Default is c("bpi_pos",
"cad_pos", "pda_pos").}

\item{headers}{Named vector mapping detector types to header names in the
mzML file.}

\item{type}{Type of chromatogram to display. Either "baselined" or
"improved". Default is "baselined".}

\item{normalize_intensity}{Normalize intensity? Default is TRUE.}

\item{normalize_time}{Normalize time? Default is FALSE.}

\item{show_example}{Show example data? Default is FALSE.}

\item{intensity_floor}{Small positive value for intensity floor. Default is
0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection".}

\item{improve_signal}{Logical. Whether to apply signal improvement (Fourier
filtering and sharpening). Default is TRUE.}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_chromatograms_alignment(show_example = TRUE)
}
}
