% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chromatograms_alignment.R
\name{check_chromatograms_alignment}
\alias{check_chromatograms_alignment}
\title{Check chromatograms alignment}
\usage{
check_chromatograms_alignment(
  file_negative = NULL,
  file_positive = NULL,
  time_min = 0.5,
  time_max = 32.5,
  cad_shift = 0.05,
  pda_shift = 0.1,
  fourier_components = 0.01,
  frequency = 1,
  resample = 1,
  chromatograms = c("bpi_pos", "cad_pos", "pda_pos"),
  type = "baselined",
  normalize_intensity = TRUE,
  normalize_time = FALSE,
  show_example = FALSE
)
}
\arguments{
\item{file_negative}{Negative file path}

\item{file_positive}{Positive file path}

\item{time_min}{Minimum time}

\item{time_max}{Maximum time}

\item{cad_shift}{CAD shift}

\item{pda_shift}{PDA shift}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{resample}{Resample}

\item{chromatograms}{Chromatograms to plot}

\item{type}{Type. "baselined" or "improved"}

\item{normalize_intensity}{Normalize intensity? Default to TRUE}

\item{normalize_time}{Normalize time? Default to FALSE}

\item{show_example}{Show example? Default to FALSE}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_chromatograms_alignment(show_example = TRUE)
}
}
