% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peaks_integration.R
\name{check_peaks_integration}
\alias{check_peaks_integration}
\title{Check chromatograms alignment}
\usage{
check_peaks_integration(
  file = NULL,
  features = NULL,
  detector = "cad",
  chromatogram = "baselined",
  min_area = 0.005,
  min_intensity = 10000,
  shift = 0.05,
  show_example = FALSE,
  fourier_components = 0.01,
  time_min = 0,
  time_max = Inf,
  frequency = 2,
  resample = 1
)
}
\arguments{
\item{file}{File path}

\item{features}{Features path}

\item{detector}{Detector}

\item{chromatogram}{Chromatogram}

\item{min_area}{Minimum area}

\item{min_intensity}{Minimum intensity}

\item{shift}{shift}

\item{show_example}{Show example? Default to FALSE}

\item{fourier_components}{Fourier components}

\item{time_min}{Time min}

\item{time_max}{Time max}

\item{frequency}{Frequency}

\item{resample}{Resample}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_peaks_integration(show_example = TRUE)
}
}
