% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peaks_integration.R
\name{check_peaks_integration}
\alias{check_peaks_integration}
\title{Check chromatograms alignment}
\usage{
check_peaks_integration(
  file = NULL,
  features = NULL,
  detector = "cad",
  chromatogram = "baselined",
  headers = c(bpi = "BasePeak_0", pda = "PDA#1_TotalAbsorbance_0", cad = "UV#1_CAD_1_0"),
  min_area = 0.005,
  min_intensity = 10000,
  shift = 0.05,
  show_example = FALSE,
  fourier_components = 0.01,
  time_min = 0.5,
  time_max = 32.5,
  frequency = 1,
  resample = 1,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection",
  sd_max = 50,
  max_iter = 1000,
  noise_threshold = 0.001,
  fit = "egh",
  intensity_threshold = 0.1,
  improve_signal = TRUE
)
}
\arguments{
\item{file}{File path}

\item{features}{Features path}

\item{detector}{Detector type (e.g., "cad", "bpi", "pda")}

\item{chromatogram}{Chromatogram type. One of "original", "improved", or
"baselined". Default is "baselined".}

\item{headers}{Named vector mapping detector types to header names.}

\item{min_area}{Minimum area fraction for peak filtering. Default is 0.005.}

\item{min_intensity}{Minimum intensity for feature filtering. Default is
1E4.}

\item{shift}{Time shift in minutes. Default is 0.05.}

\item{show_example}{Show example data? Default is FALSE.}

\item{fourier_components}{Fraction of Fourier components to keep. Default is
0.01.}

\item{time_min}{Time min in minutes. Default is 0.5.}

\item{time_max}{Time max in minutes. Default is 32.5.}

\item{frequency}{Acquisition frequency in Hz. Default is 1.}

\item{resample}{Resampling factor. Default is 1.}

\item{intensity_floor}{Small positive value for intensity floor. Default is
0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection".}

\item{sd_max}{Maximum standard deviation for peak filtering. Default is 50.}

\item{max_iter}{Maximum iterations for peak fitting. Default is 1000.}

\item{noise_threshold}{Noise threshold for peak detection. Default is 0.001.}

\item{fit}{Peak fitting method. One of "egh", "gaussian", or "raw". Default
is "egh".}

\item{intensity_threshold}{Minimum normalized intensity threshold for
filtering. Default is 0.1.}

\item{improve_signal}{Logical. Whether to apply signal improvement. Default
is TRUE.}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_peaks_integration(show_example = TRUE)
}
}
