% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peaks_integration.R
\name{check_peaks_integration}
\alias{check_peaks_integration}
\title{Check chromatograms alignment}
\usage{
check_peaks_integration(
  file = NULL,
  features = NULL,
  detector = "cad",
  chromatogram = "baselined",
  headers = c(bpi = "BasePeak_0", pda = "PDA#1_TotalAbsorbance_0", cad = "UV#1_CAD_1_0"),
  min_area = 0.005,
  min_intensity = 10000,
  shift = 0.05,
  show_example = FALSE,
  fourier_components = 0.01,
  time_min = 0.5,
  time_max = 32.5,
  frequency = 1,
  resample = 1,
  intensity_offset = 100,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection",
  sd_max = 50,
  max_iter = 1000,
  noise_threshold = 0.001,
  fit = "egh",
  intensity_threshold = 0.1
)
}
\arguments{
\item{file}{File path}

\item{features}{Features path}

\item{detector}{Detector}

\item{chromatogram}{Chromatogram}

\item{headers}{Headers}

\item{min_area}{Minimum area}

\item{min_intensity}{Minimum intensity}

\item{shift}{shift}

\item{show_example}{Show example? Default to FALSE}

\item{fourier_components}{Fourier components}

\item{time_min}{Time min}

\item{time_max}{Time max}

\item{frequency}{Frequency}

\item{resample}{Resample}

\item{intensity_offset}{Offset to add to intensity values to handle negative
intensities. Default is 100.}

\item{intensity_floor}{Small value subtracted from minimum intensity.
Default is 0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection".}

\item{sd_max}{Maximum standard deviation for peak filtering. Default is 50.}

\item{max_iter}{Maximum iterations for peak fitting. Default is 1000.}

\item{noise_threshold}{Noise threshold for peak detection. Default is 0.001.}

\item{fit}{Peak fitting method. One of "egh", "gaussian", or "raw". Default
is "egh".}

\item{intensity_threshold}{Minimum normalized intensity threshold for
filtering. Default is 0.1.}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Check chromatograms alignment
}
\examples{
\dontrun{
check_peaks_integration(show_example = TRUE)
}
}
