% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ms_progress.R
\name{extract_ms_progress}
\alias{extract_ms_progress}
\title{Extract MS progress}
\usage{
extract_ms_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of extracted MS peaks
}
\description{
Extract MS progress
}
\examples{
NULL
}
