% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate IDs}
\usage{
generate_ids(
  taxa = c("Swertia", "Kopsia", "Ginkgo"),
  comparison = c("Swertia", "Kopsia"),
  no_stereo = TRUE,
  filter_ms_conditions = TRUE,
  create_dir_f =
    "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima/main/R/create_dir.R",
  get_file_f =
    "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima/main/R/get_file.R",
  get_last_version_from_zenodo_f =
    "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima/main/R/get_last_version_from_zenodo.R",
  log_debug_f =
    "https://raw.githubusercontent.com/taxonomicallyinformedannotation/tima/main/R/log_debug.R"
)
}
\arguments{
\item{taxa}{Taxa}

\item{comparison}{Comparison}

\item{no_stereo}{No stereo}

\item{filter_ms_conditions}{Filter MS conditions}

\item{create_dir_f}{Create dir function}

\item{get_file_f}{Get file function}

\item{get_last_version_from_zenodo_f}{Get last version from Zenodo function}
}
\value{
IDs
}
\description{
Generate IDs
}
\examples{
\dontrun{
generate_ids()
}
}
