% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate IDs}
\usage{
generate_ids(
  taxa = c("Swertia", "Kopsia", "Ginkgo"),
  comparison = c("Swertia", "Kopsia"),
  no_stereo = TRUE,
  filter_ms_conditions = TRUE,
  start = "0",
  end = "9999",
  limit = "1000000"
)
}
\arguments{
\item{taxa}{Taxa}

\item{comparison}{Comparison}

\item{no_stereo}{No stereo}

\item{filter_ms_conditions}{Filter MS conditions}

\item{start}{Start}

\item{end}{End}

\item{limit}{Limit}
}
\value{
IDs
}
\description{
Generate IDs
}
\examples{
\dontrun{
generate_ids()
}
}
