% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudochromatograms.R
\name{generate_pseudochromatograms}
\alias{generate_pseudochromatograms}
\title{Generate pseudochromatograms}
\usage{
generate_pseudochromatograms(
  annotations = NULL,
  features_informed = NULL,
  features_not_informed = NULL,
  file = NULL,
  headers = c(bpi = "BasePeak_0", pda = "PDA#1_TotalAbsorbance_0", cad = "UV#1_CAD_1_0"),
  detector = "cad",
  show_example = FALSE,
  min_confidence = 0.4,
  min_similarity_prefilter = 0.6,
  min_similarity_filter = 0.8,
  mode = "pos",
  organism = "Swertia chirayita",
  fourier_components = 0.01,
  frequency = 1,
  resample = 1,
  shift = 0.05,
  time_min = 0.5,
  time_max = 32.5,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection",
  improve_signal = TRUE
)
}
\arguments{
\item{annotations}{Annotations file path}

\item{features_informed}{Features informed file path}

\item{features_not_informed}{Features not informed file path}

\item{file}{mzML file path}

\item{headers}{Named vector mapping detector types to header names.}

\item{detector}{Detector type (e.g., "cad", "bpi", "pda")}

\item{show_example}{Show example data? Default is FALSE.}

\item{min_confidence}{Minimum confidence score. Default is 0.4.}

\item{min_similarity_prefilter}{Minimum similarity for pre-filtering.
Default is 0.6.}

\item{min_similarity_filter}{Minimum similarity for final filtering. Default
is 0.8.}

\item{mode}{Ionization mode. Either "pos" or "neg". Default is "pos".}

\item{organism}{Organism name for taxonomic filtering.}

\item{fourier_components}{Fraction of Fourier components to keep. Default is
0.01.}

\item{frequency}{Acquisition frequency in Hz. Default is 1.}

\item{resample}{Resampling factor. Default is 1.}

\item{shift}{Time shift in minutes. Default is 0.05.}

\item{time_min}{Time min in minutes. Default is 0.5.}

\item{time_max}{Time max in minutes. Default is 32.5.}

\item{intensity_floor}{Small positive value for intensity floor. Default is
0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection".}

\item{improve_signal}{Logical. Whether to apply signal improvement. Default
is TRUE.}
}
\value{
A list of plots
}
\description{
Generate pseudochromatograms
}
\examples{
\dontrun{
generate_pseudochromatograms(show_example = TRUE)
}
}
