% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_signal.R
\name{improve_signal}
\alias{improve_signal}
\title{Title}
\usage{
improve_signal(
  df,
  fourier_components = 0.01,
  time_min = TIME_MIN,
  time_max = TIME_MAX,
  frequency = 2,
  resample = 1
)
}
\arguments{
\item{resample}{}
}
\description{
Title
}
