% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_signal.R
\name{improve_signal}
\alias{improve_signal}
\title{Improve signal}
\usage{
improve_signal(
  df,
  fourier_components = 0.01,
  frequency = 2,
  resample = 1,
  time_min = 0,
  time_max = Inf,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8
)
}
\arguments{
\item{df}{Dataframe with columns 'rtime' and 'intensity'}

\item{fourier_components}{Fraction of Fourier components to keep for
filtering. Default is 0.01 (1\%). Lower values provide more smoothing.}

\item{frequency}{Acquisition frequency in Hz. Default is 2.}

\item{resample}{Resampling factor. Default is 1.}

\item{time_min}{Time min in minutes. Default is 0.}

\item{time_max}{Time max in minutes. Default is Inf.}

\item{intensity_floor}{Small positive value to ensure all intensities are
strictly positive after shifting. Default is 0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}
}
\value{
A dataframe with improved signal
}
\description{
Improve signal
}
\examples{
NULL
}
