% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_signals_progress.R
\name{improve_signals_progress}
\alias{improve_signals_progress}
\title{Improve signals progress}
\usage{
improve_signals_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of dataframes with improved signals
}
\description{
Improve signals progress
}
\examples{
NULL
}
