% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromatograms.R
\name{load_chromatograms}
\alias{load_chromatograms}
\title{Load chromatograms}
\usage{
load_chromatograms(file = NULL, show_example = FALSE, example_polarity = "pos")
}
\arguments{
\item{file}{File}

\item{show_example}{Show example? Default to FALSE}

\item{example_polarity}{Example polarity}
}
\value{
A list of chromatograms
}
\description{
Load chromatograms
}
\examples{
NULL
}
