% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_chromatograms_list.R
\name{normalize_chromatograms_list}
\alias{normalize_chromatograms_list}
\title{Normalize chromatograms list}
\usage{
normalize_chromatograms_list(
  list,
  shift = 0,
  normalize_intensity = TRUE,
  normalize_time = FALSE
)
}
\arguments{
\item{list}{List}

\item{shift}{Shift}

\item{normalize_intensity}{Normalize time}

\item{normalize_time}{Normalize intensity}
}
\value{
A dataframe with normalized chromatograms
}
\description{
Normalize chromatograms list
}
\examples{
NULL
}
