% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms_taxo}
\alias{plot_histograms_taxo}
\title{Plot histograms taxo}
\usage{
plot_histograms_taxo(dataframe, level = "max", mode = "pos")
}
\arguments{
\item{dataframe}{Dataframe}

\item{level}{Level}

\item{mode}{Mode}
}
\value{
A plot of taxo histograms
}
\description{
Plot histograms taxo
}
\examples{
NULL
}
