% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results_1}
\alias{plot_results_1}
\title{Plot results 1}
\usage{
plot_results_1(detector = "cad", mode = "pos")
}
\arguments{
\item{detector}{Detector}

\item{mode}{Mode}
}
\value{
A list of plots
}
\description{
Plot results 1
}
\examples{
NULL
}
