% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results_1}
\alias{plot_results_1}
\title{Title}
\usage{
plot_results_1(detector = "cad", mode = "pos")
}
\arguments{
\item{mode}{}
}
\description{
Title
}
\examples{
TODO harmonize 'others' among minor and major
}
