% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comparison.R
\name{prepare_comparison}
\alias{prepare_comparison}
\title{Prepare comparison}
\usage{
prepare_comparison(detector = "cad")
}
\arguments{
\item{detector}{Detector}
}
\value{
A list of peaks
}
\description{
Prepare comparison
}
\examples{
NULL
}
