% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comparison.R
\name{prepare_comparison}
\alias{prepare_comparison}
\title{Prepare comparison}
\usage{
prepare_comparison(
  features_informed = NULL,
  features_not_informed = NULL,
  min_similarity_prefilter = 0.6,
  min_similarity_filter = 0.8,
  mode = "pos"
)
}
\arguments{
\item{features_informed}{Features informed}

\item{features_not_informed}{Features not informed}

\item{min_similarity_prefilter}{Min similarity pre filter}

\item{min_similarity_filter}{Min similarity filter}

\item{mode}{Mode}
}
\value{
A list of peaks
}
\description{
Prepare comparison
}
\examples{
NULL
}
