% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy.R
\name{prepare_hierarchy}
\alias{prepare_hierarchy}
\title{Title}
\usage{
prepare_hierarchy(
  dataframe,
  type = "analysis",
  detector = "ms",
  rescale = FALSE
)
}
\arguments{
\item{rescale}{}
}
\description{
Title
}
