% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_preparation.R
\name{prepare_hierarchy_preparation}
\alias{prepare_hierarchy_preparation}
\title{Prepare hierarchy preparation}
\usage{
prepare_hierarchy_preparation(dataframe)
}
\arguments{
\item{dataframe}{Dataframe}
}
\value{
A dataframe with a description todo
}
\description{
Prepare hierarchy preparation
}
\examples{
NULL
}
