% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_chromatograms.R
\name{preprocess_chromatograms}
\alias{preprocess_chromatograms}
\title{Preprocess chromatograms}
\usage{
preprocess_chromatograms(
  detector = "cad",
  fourier_components = 0.01,
  frequency = 2,
  list,
  name,
  resample = 1,
  shift = 0,
  time_min = 0,
  time_max = Inf,
  intensity_offset = 100,
  intensity_floor = 0.001,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  smoothing_width = 8,
  baseline_method = "peakDetection"
)
}
\arguments{
\item{detector}{Detector}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{list}{List}

\item{name}{Name}

\item{resample}{Resample}

\item{shift}{Shift}

\item{time_min}{Time min}

\item{time_max}{Time max}

\item{intensity_offset}{Offset to add to intensity values to handle negative
intensities. Default is 100.}

\item{intensity_floor}{Small value subtracted from minimum intensity.
Default is 0.001.}

\item{k2}{K2 parameter for signal sharpening. Default is 250.}

\item{k4}{K4 parameter for signal sharpening. Default is 1250000.}

\item{sigma}{Sigma parameter for signal sharpening. Default is 0.05.}

\item{smoothing_width}{Smoothing width for signal sharpening. Default is 8.}

\item{baseline_method}{Method for baseline correction. Default is
"peakDetection". See \code{\link[baseline]{baseline}} for available
methods.}
}
\value{
A list of preprocessed chromatograms
}
\description{
Preprocess chromatograms
}
\examples{
NULL
}
