% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_chromatograms.R
\name{preprocess_chromatograms}
\alias{preprocess_chromatograms}
\title{Preprocess chromatograms}
\usage{
preprocess_chromatograms(
  detector = "cad",
  fourier_components = 0.01,
  frequency = 2,
  list,
  name,
  resample = 1,
  shift = 0,
  time_min = 0,
  time_max = Inf
)
}
\arguments{
\item{detector}{Detector}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{list}{List}

\item{name}{Name}

\item{resample}{Resample}

\item{shift}{Shift}

\item{time_min}{Time min}

\item{time_max}{Time max}

\item{signal_name}{Signal name}
}
\value{
A list of preprocessed chromatograms
}
\description{
Preprocess chromatograms
}
\examples{
NULL
}
