% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_chromatograms.R
\name{preprocess_chromatograms}
\alias{preprocess_chromatograms}
\title{Preprocess chromatograms}
\usage{
preprocess_chromatograms(
  detector = "cad",
  list = chromatograms_all[c(FALSE, FALSE, TRUE)],
  signal_name = "UV.1_CAD_1_0",
  shift = CAD_SHIFT
)
}
\arguments{
\item{detector}{Detector}

\item{list}{List}

\item{signal_name}{Signal name}

\item{shift}{Shift}
}
\value{
A list of preprocessed chromatograms
}
\description{
Preprocess chromatograms
}
\examples{
NULL
}
