% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_peaks.R
\name{preprocess_peaks}
\alias{preprocess_peaks}
\title{Preprocess peaks}
\usage{
preprocess_peaks(
  detector = "cad",
  list = chromatograms_list_cad$chromatograms_baselined,
  name,
  df_long = chromatograms_list_cad$chromatograms_baselined_long,
  area_min = 0
)
}
\arguments{
\item{detector}{Detector}

\item{list}{List}

\item{name}{Name}

\item{df_long}{DF long}

\item{area_min}{Area min}
}
\value{
A list of lists and dataframe with preprocessed peaks
}
\description{
Preprocess peaks
}
\examples{
NULL
}
