% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_peaks.R
\name{preprocess_peaks}
\alias{preprocess_peaks}
\title{Title}
\usage{
preprocess_peaks(
  detector = "cad",
  list = chromatograms_list_cad$chromatograms_improved,
  df_long = chromatograms_list_cad$chromatograms_improved_long
)
}
\arguments{
\item{detector}{}
}
\description{
Title
}
\examples{
data.table call outside of future because buggy else
data.table call outside of future because buggy else
}
