% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_compare_peaks.R
\name{process_compare_peaks}
\alias{process_compare_peaks}
\title{Process compare peaks}
\usage{
process_compare_peaks(
  file = NULL,
  features = NULL,
  type = "baselined",
  detector = "cad",
  export_dir = "data/interim/peaks",
  show_example = FALSE,
  fourier_components = 0.01,
  frequency = 2,
  min_area = 0.005,
  min_intensity = 10000,
  resample = 1,
  shift = 0.05,
  time_min = 0,
  time_max = Inf
)
}
\arguments{
\item{file}{File path}

\item{features}{Features path}

\item{type}{Type. "original", "baselined" or "improved"}

\item{detector}{Detector}

\item{export_dir}{Export directory}

\item{show_example}{Show example? Default to FALSE}

\item{fourier_components}{Fourier components}

\item{frequency}{Frequency}

\item{min_area}{Min area}

\item{min_intensity}{Min intensity}

\item{resample}{Resample}

\item{shift}{Shift}

\item{time_min}{Time min}

\item{time_max}{Time max}
}
\value{
A plot with (non-)aligned chromatograms
}
\description{
Process compare peaks
}
\examples{
\dontrun{
check_chromatograms_alignment(show_example = TRUE)
}
}
