% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_progress.R
\name{queries_progress}
\alias{queries_progress}
\title{Queries progress}
\usage{
queries_progress(xs)
}
\arguments{
\item{xs}{XS}
}
\value{
A list of queries results
}
\description{
Queries progress
}
\examples{
NULL
}
