% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_progress.R
\name{queries_progress}
\alias{queries_progress}
\title{Queries progress}
\usage{
queries_progress(xs, start = "0", end = "9999", limit = "1000000")
}
\arguments{
\item{xs}{XS}

\item{start}{Start}

\item{end}{End}

\item{limit}{Limit}
}
\value{
A list of queries
}
\description{
Queries progress
}
\examples{
NULL
}
