% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCADE.R
\name{sc_ECO}
\alias{sc_ECO}
\title{Ecological (Eco) score calculation for single cell samples.
Yields a hazard score, resource score, and ecological score for each sample.}
\usage{
sc_ECO(tme_comp, subtype_scores)
}
\arguments{
\item{tme_comp}{Data frame containing nontumor cell subtype composition for
each sample. Samples are on the rows, subtypes on the columns. Rows must be
scaled to sum to 1. Column names (subtypes) must all be included in the
rownames of \code{subtype_scores}}

\item{subtype_scores}{Data frame containing nontumor cell subtype weights
and labels. Rownames are subtypes. Must contain a \code{category} column that
labels each subtype as a "Hazard" or "Resource" and a \code{score} column that
lists a weight for each subtype. Defaults to a built-in weight table curated
from a single-cell study of liver cancer.}
}
\value{
Data frame with three columns: "hazards" - the hazard score,
"resources" - the resource score, and "eco_score" - the ecological score
for each sample. Samples are row names.
}
\description{
Hazard score describes the amount of hazards facing a tumor.
Resource score describes the amount of resources available to a tumor.
Eco score describes the overall state of the TME with respect to a tumor,
where a low score indicates hazardous conditions and a high score indicates
resource-heavy conditions. This metric is uniformly distributed from 0 to 1.
}
\details{
Hazards and resources are described from the perspective of a tumor, where
hazards attack it while resources aid its survival.
}
