% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCADE.R
\name{sc_LINEAGE}
\alias{sc_LINEAGE}
\title{Lineage score calculation for single cell samples}
\usage{
sc_LINEAGE(expr_data, idents, lineage_genes)
}
\arguments{
\item{expr_data}{Matrix of normalized gene counts from single-cell dataset.
Genes are on the rows and samples are on the columns.}

\item{idents}{Vector of sample labels for the columns of \code{expr_data}.}

\item{lineage_genes}{List of vectors containing genes for each lineage.
Gene set vectors must be named after the lineages that they represent.
Defaults to a built-in list of genes curated from a single-cell study of
liver cancer. Contains gene sets for CLTCs (cholangiocyte-like tumor cells),
HLTCs (hepatocyte-like tumor cells), MCTCs (myeloid cell-like tumor cells),
and MLTCs (mesenchymal-like tumor cells). If an independent value is given
for this parameter, it must contain all 4 of these lineages in order to
calculate a lineage score.}
}
\value{
Data frame containing average gene expression for each lineage and
the final lineage score in the "lineage_score" column. Samples are row names.
}
\description{
Lineage score calculation for single cell samples
}
