% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sharpening.R
\name{signal_sharpening}
\alias{signal_sharpening}
\title{Signal sharpening}
\usage{
signal_sharpening(
  time,
  intensity,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  Smoothing_width = 8,
  Baseline_adjust = 0
)
}
\arguments{
\item{time}{time}

\item{intensity}{intensity}

\item{k2}{K2}

\item{k4}{K4}

\item{sigma}{Sigma}

\item{Smoothing_width}{Smoothing width}

\item{Baseline_adjust}{Baseline adjust}
}
\value{
A sharpened signal
}
\description{
Signal sharpening
}
\examples{
NULL
}
