% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sharpening.R
\name{signal_sharpening}
\alias{signal_sharpening}
\title{Signal sharpening}
\usage{
signal_sharpening(
  time,
  intensity,
  k2 = 250,
  k4 = 1250000,
  sigma = 0.05,
  Smoothing_width = 8,
  Baseline_adjust = 0
)
}
\arguments{
\item{time}{time}

\item{intensity}{intensity}

\item{k2}{K2 parameter controlling the weight of the second derivative in
signal sharpening. Default is 250. Lower values increase the sharpening
effect from the second derivative.}

\item{k4}{K4 parameter controlling the weight of the fourth derivative in
signal sharpening. Default is 1250000. Lower values increase the
sharpening effect from the fourth derivative.}

\item{sigma}{Sigma parameter for derivative weighting. Default is 0.05.
Higher values increase the overall sharpening effect.}

\item{Smoothing_width}{Smoothing width for the running mean filter. Default
is 8. Higher values provide more smoothing but reduce resolution.}

\item{Baseline_adjust}{Baseline adjustment value. Default is 0.}
}
\value{
A sharpened signal
}
\description{
Signal sharpening
}
\examples{
NULL
}
