% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sharpening.R
\name{signal_sharpening}
\alias{signal_sharpening}
\title{Title}
\usage{
signal_sharpening(
  Time = timeow,
  Intensity = intensityeah,
  K2 = k2,
  K4 = k4,
  Smoothing_width = smoothing_width,
  Baseline_adjust = baseline_adjust
)
}
\arguments{
\item{Baseline_adjust}{}
}
\description{
Title
}
