% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_progress.R
\name{tables_progress}
\alias{tables_progress}
\title{Tables progress}
\usage{
tables_progress(xs, structures_classified)
}
\arguments{
\item{xs}{XS}

\item{structures_classified}{structures classified}
}
\value{
A list of tables
}
\description{
Tables progress
}
\examples{
NULL
}
